Refactor tests organization: one AbstractTestCase extended into AbstractTestCaseIidm and AbstractTestCaseNoIidm Separation of iidm and noIidm test into different packages.
