+++ single-line-diagram-core/src/test/java/com/powsybl/sld/raw/AbstractTestCaseRaw.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.noIidm;
+package com.powsybl.sld.raw;
 
 import com.powsybl.sld.AbstractTestCase;
 import com.powsybl.sld.RawGraphBuilder;
@@ -12,7 +12,7 @@ import com.powsybl.sld.RawGraphBuilder;
 /**
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
  */
-public abstract class AbstractTestCaseNoIidm extends AbstractTestCase {
+public abstract class AbstractTestCaseRaw extends AbstractTestCase {
 
     protected RawGraphBuilder rawGraphBuilder = new RawGraphBuilder();
