+++ single-line-diagram-core/src/test/java/com/powsybl/sld/raw/TestCase6.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.noIidm;
+package com.powsybl.sld.raw;
 
 import com.powsybl.sld.layout.BlockOrganizer;
 import com.powsybl.sld.layout.ImplicitCellDetector;
@@ -18,42 +18,42 @@ import static org.junit.Assert.assertEquals;
 
 /**
  * <pre>
- *
- *       la     lb
- *       |      |
- *      nsa-bs-nsb
- *       |      |
- *       ba     bb
- *       |      |
- * bbs---da-----db---
+ *              b
+ *           /     \
+ *          |       |
+ * bbs1.1 -d1- ds1 -|-- bbs1.2
+ * bbs2.1 ---- ds2 -d2- bbs2.2
  *
  * </pre>
  *
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
  */
-public class TestCase5H extends AbstractTestCaseNoIidm {
+
+public class TestCase6 extends AbstractTestCaseRaw {
 
     @Before
     public void setUp() {
         com.powsybl.sld.RawGraphBuilder.VoltageLevelBuilder vlBuilder = rawGraphBuilder.createVoltageLevelBuilder("vl", 400);
-        BusNode bbs = vlBuilder.createBusBarSection("bbs", 1, 1);
-        FeederNode la = vlBuilder.createLoad("la", 10, BusCell.Direction.TOP);
-        SwitchNode ba = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.BREAKER, "ba", false, false);
-        SwitchNode da = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.DISCONNECTOR, "da", false, false);
-        vlBuilder.connectNode(la, ba);
-        vlBuilder.connectNode(ba, da);
-        vlBuilder.connectNode(da, bbs);
 
-        FeederNode lb = vlBuilder.createLoad("lb", 20, BusCell.Direction.TOP);
-        SwitchNode bb = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.BREAKER, "bb", false, false);
-        SwitchNode db = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.DISCONNECTOR, "db", false, false);
-        vlBuilder.connectNode(lb, bb);
-        vlBuilder.connectNode(bb, db);
-        vlBuilder.connectNode(db, bbs);
+        BusNode bbs11 = vlBuilder.createBusBarSection("bbs1.1", 1, 1);
+        BusNode bbs12 = vlBuilder.createBusBarSection("bbs1.2", 1, 2);
+        BusNode bbs21 = vlBuilder.createBusBarSection("bbs2.1", 2, 1);
+        BusNode bbs22 = vlBuilder.createBusBarSection("bbs2.2", 2, 2);
 
-        SwitchNode bs = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.BREAKER, "bs", false, false);
-        vlBuilder.connectNode(la, bs);
-        vlBuilder.connectNode(lb, bs);
+        SwitchNode d1 = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.DISCONNECTOR, "d1", false, false);
+        SwitchNode d2 = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.DISCONNECTOR, "d2", false, false);
+        SwitchNode b = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.BREAKER, "b", false, false);
+        vlBuilder.connectNode(bbs11, d1);
+        vlBuilder.connectNode(d1, b);
+        vlBuilder.connectNode(d2, b);
+        vlBuilder.connectNode(d2, bbs22);
+
+        SwitchNode ds1 = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.DISCONNECTOR, "ds1", false, false);
+        vlBuilder.connectNode(bbs11, ds1);
+        vlBuilder.connectNode(bbs12, ds1);
+        SwitchNode ds2 = vlBuilder.createSwitchNode(SwitchNode.SwitchKind.DISCONNECTOR, "ds2", false, false);
+        vlBuilder.connectNode(bbs21, ds2);
+        vlBuilder.connectNode(bbs22, ds2);
 
     }
 
@@ -89,6 +89,6 @@ public class TestCase5H extends AbstractTestCaseNoIidm {
         new PositionVoltageLevelLayout(g).run(layoutParameters);
 
         // write Json and compare to reference
-        assertEquals(toString("/TestCase5ShuntHorizontal.json"), toJson(g, "/TestCase5H.json"));
+        assertEquals(toString("/TestCase6CouplingNonFlatHorizontal.json"), toJson(g, "/TestCase6.json"));
     }
 }
